refactor: separate Moment from SingleCardLayout
